[custom_ops] make custom_ops.impl work on existing operators
The design is that we construct a CustomOp object around the existing
operator and then use it to register things. It is totally OK if the
operator isn't functional (unlike torch._custom_ops.custom_op that can
only construct functional operators).

If the operator already has an implementation from a backend (either via
direct registration to e.g. DispatchKey::CPU, or an indirect
registration like CompositeImplicitAutograd/CompositeExplicitAutograd),
we raise an error.

Test Plan:
- new tests

ghstack-source-id: 3f9a4eedd35b29b0258746aae68f6acf170962da
Pull Request resolved: https://github.com/pytorch/pytorch/pull/106076