[Dynamo] Cache {NNModuleSource, AttrSource}.{guard_source, name}
The string building done with these two methods accounted for 50% (2m)
of a model I was looking at. Caching these removed 50% of the
tracing time.

NB: this is a short-term fix. Horace and Voz pointed out that these
come from us cloning VariableTrackers and re-inserting their guards. A
better long-term state is to refactor that logic.

Test Plan:
- existing tests