Do not import sympy within torch._prims_common
This is the first of a few PRs that avoid importing SymPy at import time.
The pitch here is that we (almost!) do not have SymPy on our API, so
this should be feasible.

This should speed-up torch imports by a good 15% as per
https://dev-discuss.pytorch.org/t/delving-into-what-happens-when-you-import-torch/1589

In this PR we just move a few global imports into local imports.

ghstack-source-id: 92e1e2543096bff5c3c4fe9e791b2ce1a52405ed
Pull Request resolved: https://github.com/pytorch/pytorch/pull/112034