[dynamo] Initialize tensor_weakref_to_sizes_strides with a weak dict
Spotted while working on getting output_graph.py to typecheck.

The type hint indicates that it was intended to be initialized with a
WeakIdKeyDictionary, but the actual runtime value was a regular dict.
Not sure if there's some kind of test we should add for this fix.

Looks like the code was originally added in
https://github.com/pytorch/pytorch/pull/100128.

ghstack-source-id: a8fb35258968fd6c8ef36d4bbf9f85cdd5ffc3c1
Pull Request resolved: https://github.com/pytorch/pytorch/pull/113412