Make variables in dict lazy and remove DICT_KEYS guard
We build the keys of a dict depth-first and we rely on the guards of
each element in the dict to create the correct guards. This allows us to
remove the rather buggy DICT_KEYS guard and make the guard lazy.

Also, adding new types to the supported types within keys should be less
error prone.

This is marginally less efficient when we graph break, but in turn we
should graph break much less. It also  makes the dicts code easier to maintain
(removes `is_hashable_python_var`).

ghstack-source-id: 591f5d57905e83c03a4785a2c34168221c8e6d8a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/117625