Fix test failure: Add torch.cuda._get_device_properties to dynamo trace rules
In this PR stack, there were unrelated test failures within test_trace_rules.py - It turned out that torch.cuda._get_device_properties should be registered in _dynamoc/trace_rules.py. A test failed because it was not.

This is a small fix which tries to get rid of the test failure by manually registering that function.

Note:
I am not sure whether this is the best way to fix this, as I am neither familiar with the trace rules nor with the introduction of torch.cuda._get_device_properties.

ghstack-source-id: 01afda2258c679f2365fe40634f2af08a281b4dc
Pull Request resolved: https://github.com/pytorch/pytorch/pull/120620