CI: Download triton dependencies once between all wheel build jobs
The LLVM download in particular is slow and is hitting timeouts in CI before the
compiler is even called once. I workaround this by moving the download step into
its own job with a 1 hour timeout and in the other jobs we download from an
artifact instead of the download server.

This reduces strain on the download server and gives us a more generous timeout.

ghstack-source-id: 546d44ec5567206e9d1f06138d69fff365a90ade
Pull Request resolved: https://github.com/pytorch/pytorch/pull/121568