CI: Download triton dependencies once between all wheel build jobs
The LLVM download in particular is slow and is hitting timeouts in CI before the
compiler is even called once. I workaround this by moving the download step into
its own job with a 1 hour timeout and in the other jobs we download from an
artifact instead of the download server.

This reduces strain on the download server and gives us a more generous timeout.

ghstack-source-id: 293ca3c2107eac0e52c006c2137ea6b8d739c982
Pull Request resolved: https://github.com/pytorch/pytorch/pull/121568