[inductor] FX graph cache: simplify "current callable" logic
Summary: The handling of the current_callable and compiled_artifact fields in the CompiledFxGraph object is unnecessarily complicated and confusing. We can simplify by storing only the callable. That field is not serializable, so the caching approach is to store a path to the generated artifact and reload from disk on a cache hit. We can just reload inline in the FX cache hit path. This change has the added benefit that it makes it easier to fallback to a "cache miss" if the path somehow doesn't exist.

ghstack-source-id: 34a0a87f58df45d3d130eab6d1d0b1c852b10225
Pull Request resolved: https://github.com/pytorch/pytorch/pull/121903