triton_kernel_wrap shouldn't access FakeTensor.data_ptr
The comment suggests that we need to replace all FakeTensors with real
tensors. `torch.empty` doesn't actually return a real Tensor because
FakeTensorMode is active!

We disable torch dispatch so that torch.empty actually returns a real Tensor.

The motivation for this PR is that we're trying to ban
FakeTensor.data_ptr (or at least warn on it) in torch.compile. See the
next PR up in the stack

Test Plan:
- Existing tests.

ghstack-source-id: 861612bb9fa1ba7bbada00cbc3fc959d24570568
Pull Request resolved: https://github.com/pytorch/pytorch/pull/122418