Allow fake models to run with ONNXProgram.__call__
In order to a fake model to run using ONNXProgram.__call__
interface, we need to save the model into disk along with external data
before executing the model. This is what this PR implements

An alternative is to ONNXProgram.__call__ to detect that the model
was exported with fake mode and explicit raise an exception when
ONNXProgram.__call__ is executed. The exception message would instruct
the user to call ONNXProgram.save and manually execute the model using
the ONNX runtime of choice.

ghstack-source-id: 861b78cb08614aa8b2539c09bb320a9badc694bc
Pull Request resolved: https://github.com/pytorch/pytorch/pull/122230