Pop codegened parent graph from wrapper in GraphLowering
Summary: Previously, we kept a reference to `V.graph` in the `codegened_graph_stack` of the wrapper. Memory regression analysis of https://github.com/pytorch/pytorch/issues/121887 shows that this has led to a slightly higher memory utilization during lowering of the `llama_v2_7b_16h` model. Here we refactor the code to pop the parent subgraph from the `codegened_graph_stack` when codegen-ing is done.

Test Plan: CI

ghstack-source-id: 96b24eef31d0047557e7ea0e2ab3358be81222aa
Pull Request resolved: https://github.com/pytorch/pytorch/pull/122469