Don't create world pg variable out of thin air when rewriting c10d collectives
Previously, when rewriting c10d collectives, if the group argument is
unspecified or None, we create a world pg variable out of thin air and
pass it to the rewrite target. The approach was problematic, as it
assumes the symbol `torch` is available in the scope (see #122404).

After #120560, dynamo can now trace dist.group.WORLD. If the group
argument is unspecified, we can just set it with dist.group.WORLD in the
rewrite target.

Testing

pytest test/distributed/test_inductor_collectives.py -k test_dynamo_rewrite_dist_allreduce

Also verified with the repro provided in #122404

ghstack-source-id: 3d144bfd9c607cec1cd5468d1f6afd24a73b41dd
Pull Request resolved: https://github.com/pytorch/pytorch/pull/122561