[RFC] delete graph_out_{n} after restoring local vars
At graph breaks, we create a graph_out_{n} symbol to hold the graph output and
use it to restore the local vars. In addition to their own symbols, the local
vars are kept alive by the symbol we created. This means that if the graph
break is the last usage of one of the symbols, the symbol would still be kept
alive upon graph resumption.

This PR: delete the graph_out_{n} symbol after restoring local vars so the
lifetime of the local vars is governed by themselves.

ghstack-source-id: 4abd6a9c503b88635420271918d49fcd8ce4aadc
Pull Request resolved: https://github.com/pytorch/pytorch/pull/122658