[dynamo] Skip leaf check on `assert_metadata_eq` if grad tensor level is `-2`
When fakifying a grad tracking tensor, if the level is -2 (sentinel
value) we can just unwrap the grad tensor and return a fake version of
it. In this PR, we update the `assert_metadata_eq` to not compare if
the grad tensor and the unwrapped ones are leafs or not, as this may
not be always true.

ghstack-source-id: 694f0ccd9f8c0a21b4d3c46734756c886c9d0baa
Pull Request resolved: https://github.com/pytorch/pytorch/pull/122728