Add clone if output is a view from constant.
Summary:
For the original clone we did for output, we only clone when the
corresponding tensor is an constant. We need this because we have to
make sure the constants' ownership maintain in the Model. However we
haven't include if it's a view of a constant.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: f02fae69544d972ce99a5643fa0ed55e50fc1147
Pull Request resolved: https://github.com/pytorch/pytorch/pull/123182