Add clone if output is a view from constant.
Summary:
For the original clone we did for output, we only clone when the
corresponding tensor is an constant. We need this because we have to
make sure the constants' ownership maintain in the Model. However we
haven't include if it's a view of a constant.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: a978749997d297551949a38b97fcaf397b6e6e1c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/123185