[DCP] Implements DCP logger
Pull Request resolved: https://github.com/pytorch/pytorch/pull/121352

A large motivator for this work is unification with model store components. To that extent, please note any existing model store logs should not be changing. In the "Unified Checkpointer", existing model store logs should be be migrated either to one of the plugins passed to DCP API's (StorageReader/Writer, Save/LoadPlanner), or to the trainer level.
ghstack-source-id: 221130692
@exported-using-ghexport

Differential Revision: [D54512626](https://our.internmc.facebook.com/intern/diff/D54512626/)