Clone mutated inputs in first pass of CPP wrapper compilation
Summary: CPP wrapper compilation is currently done in two passes: in the first pass, Python wrapper is generated and run to compile Triton kernels as a side effect, in the second pass C++ wrapper is generated and compiled. When model inputs are mutated, running the Python wrapper in the first pass mutates the inputs although the entire first pass is strictly a part of the compilation process, hence must not introduce any side effects.

In this PR, we clone mutated inputs in the first pass to avoid mutation side effects.

Fixes https://github.com/pytorch/pytorch/issues/117364.

Test Plan:

```
$ TORCHINDUCTOR_CPP_WRAPPER=1 python test/inductor/test_torchinductor.py -k test_inductor_layout_optimization
_input_mutations_cuda
...
.
----------------------------------------------------------------------
Ran 1 test in 6.368s

OK
```

ghstack-source-id: 486abd77ceaa60601ace1f34c74c76b019ff82f2
Pull Request resolved: https://github.com/pytorch/pytorch/pull/123316