Add aoti_torch_dtype<> specializations for half and bfloat16
Fixes #122989. (Note that while the missing symbol issue is fixed, the
test itself is still disabled, because the test runner now segfaults on
`atexit()`; but I think that issue is unrelated to the missing symbol.)

In addition to defining the specializations, I also `= delete`d the
default un-specialized version of `aoti_torch_dtype`, so future missing
dtype references will show up as compile-time instead of link-time
errors.

ghstack-source-id: edfc416f40483ec3474279486d2a1b89248c823c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/123692