[multi_tensor_apply] revert the optimization introduced in #119764
The optimization causes regression in some torchbench benchmarks and
with some older versions of nvcc. The regression is preventable, but it
might require additional template specialization which would increase the
binary size.

Reverting it for now to re-evaluate.

Keeping the introduced tests and cuda-to-hip-mappings since these are
not specific to the optimization.

ghstack-source-id: e32eb3cc89d4ab53d3c6bee55246bd0bca690597
Pull Request resolved: https://github.com/pytorch/pytorch/pull/123763