Allow recursive calls to ExportProgramDeserializer
Summary:
In ExecuTorch there we'd like to call ExportProgramSerializer/Deserialiazer recursively to be able to serialize/deserialize delegated modules. Currently the existing logic prevents this from happening as we assert that _CURRENT_DESERIALIZER is None.

The alternative is that we convert `_CURRENT_DESERIALIZER` to a list of deserializers and check that the state of this list is the same as it was when we entered it. It should effectively be a no-op for any use cases that don't need to call this recursively.

Test Plan: CI

Differential Revision: D55996990